has_child boolean added

+++ src/signals/signals.c
@@ -1,22 +1,21 @@
 #include "signals.h"
 
-pid_t pid = -1;
-
 void handler_ctrlc_parent(int sig) {
 	signal(SIGINT, handler_ctrlc_parent); // reset 
 	
 	// recebe o signal
 	printf("Parent recebeu o signal %u do ctrl-c\n", sig);
 	
-	if (pid != -1 && pid != 0) { // HAS A CHILD??
-		// se tiver filhos!  - envia para o grupo dos filhos?
-		if (kill(pid, SIGUSR1) != 0) // TERMINAR OS PROCESSOS DE FILHOS/NETOS TODOS 
+	if (has_child) {
+		// se tiver filhos!  - envia para o grupo do filho		
+		if (kill(0, SIGUSR1) != 0) // ENVIAR PARA TODOS OS PROCESSOS COM IGUAL GROUP ID - descendentes + ele proprio
 			perror("Error sending SIGUSR1 to children!");
 	} else {
 		// senao tiver filhos! - envia SIGUSR2 para si proprio
 		kill(getpid(), SIGUSR2);
 	}
 }
+
 
 void display_child_info(int sig) {
 	signal(SIGUSR1, display_child_info); // reset 
@@ -44,20 +43,31 @@ void make_question(int sig) {
 	
 	// --> faz a pergunta e recebe a resposta
 	printf("\nPretende mesmo terminar o programa?(y/n)\n");
-	scanf("%c", &c); // just to test purposes
+    
+	scanf("%c", &c);
+    flush_in();
 	
 	if (c == 'y') { 
 		// really wants to terminate
-		if (kill(pid, SIGCONT) != 0) // TERMINAR OS PROCESSOS DE FILHOS/NETOS TODOS 
-			perror("Error sending SIGCONT to the descendents");
-		if (kill(pid, SIGTERM) != 0) // TERMINAR OS PROCESSOS DE FILHOS/NETOS TODOS 
-			perror("Error sending SIGTERM to the descendents");
-		if (kill(getpid(), SIGTERM) != 0)
-			perror("Error sending SIGTERM to him self"); // TERMINAR ELE PROPRIO
+		
+		// terminar descendentes
+		if (has_child) {
+			if (kill(0, SIGCONT) != 0) // CONTINUAR TODOS OS PROCESSOS
+				perror("Error sending SIGCONT to the descendents");
+			if (kill(0, SIGTERM) != 0) // TERMINAR TODOS PROCESSOS
+				perror("Error sending SIGTERM to the descendents");
+		}
+        else 
+            if (kill(getpid(), SIGTERM) != 0)
+			    perror("Error sending SIGTERM to him self"); // TERMINAR ELE PROPRIO
+
 	} else { 
-		// continue processing
-		if (kill(pid, SIGCONT) != 0) // TERMINAR OS PROCESSOS DE FILHOS/NETOS TODOS 
-			perror("Error sending SIGCONT to the child");
+		// continue processing descendentes
+		if (has_child) {
+			if (kill(0, SIGCONT) != 0) // CONTINUAR TODOS OS PROCESSOS
+				perror("Error sending SIGCONT to the child");
+		}
+        printf("Resuming...\n");
 	}
 }
 
@@ -65,4 +75,9 @@ void make_question(int sig) {
 void handler_term(int sig) {
 	printf("Termitating!!...\n"); // to registrate the signal
 	exit(EXIT_CTRLC); 
+}
+
+void flush_in(){ 
+    int ch;
+    while( (ch = fgetc(stdin)) != EOF && ch != '\n' ){} 
 }
